+++ lib/packages/http/src/httpClientFactory.ts
@@ -1,0 +1,8 @@
+import {Http} from './http';
+import {HttpAdapterAxios} from './httpAdapterAxios';
+
+export class HttpClientFactory {
+    static createHttpClient(baseUrl: string, options?: any): Http {
+        return new HttpAdapterAxios(baseUrl, options);
+    }
+}