+++ lib/packages/http/CHANGELOG.md
@@ -2,7 +2,9 @@
 All notable changes to this project will be documented in this file.
 
 # 0.6.0
-- Exporting HttpImpl as HttpClient
+Breaking Change
+---------------------
+- Access to Http client through new `HttpClientFactory`
 
 # 0.5.1
 - Better message handling in case of errors